Updated Admin System: Moved Vet and Clinic List to their own Panels and cleared Home Panel
